requirements.txt
discord.py
google-generativeai
yt-dlp
PyNaCl
Flask
aiohttp